LLMで生成したNPCの返答をDB連携
