+++ src/server/FastAPI/supabase_client.py
@@ -1,0 +1,10 @@
+import os
+from supabase import create_client, Client
+from dotenv import load_dotenv
+
+load_dotenv()
+
+SUPABASE_URL = os.environ.get("REACT_APP_SUPABASE_URL")
+SUPABASE_KEY = os.environ.get("REACT_APP_SUPABASE_KEY")
+
+supabase: Client = create_client(SUPABASE_URL, SUPABASE_KEY)
